+++ src/main/java/ui/widgets/buttons/SimpleButton.java
@@ -77,9 +77,9 @@ public class SimpleButton<T extends SimpleButton<T>> extends AbstractButton<T> {
     @Override public float getPrefHeight() { return height; }
 
     public void renderBackground(SpriteBatch sb, float bottomLeftX, float bottomLeftY, float width, float height) {
-        RenderingHelper.renderDynamicBase(sb, TEX_CORNER_BASE, bottomLeftX, bottomLeftY, width, height, CORNER_SIZE, getBaseColor());
-        RenderingHelper.renderDynamicHighlight(sb, TEX_CORNER_HIGHLIGHT, TEX_TOP_EDGE_HIGHLIGHT, TEX_CENTER_HIGHLIGHT, bottomLeftX, bottomLeftY, width, height, CORNER_SIZE);
-        RenderingHelper.renderDynamicTrim(sb, TEX_CORNER_TRIM, TEX_EDGE_TRIM, bottomLeftX, bottomLeftY, width, height, CORNER_SIZE, getTrimColor());
+        RenderingHelper.renderDynamicBase(sb, TEX_CORNER_BASE, (int)bottomLeftX, (int)bottomLeftY, (int)width, (int)height, CORNER_SIZE, getBaseColor());
+        RenderingHelper.renderDynamicHighlight(sb, TEX_CORNER_HIGHLIGHT, TEX_TOP_EDGE_HIGHLIGHT, TEX_CENTER_HIGHLIGHT, (int)bottomLeftX, (int)bottomLeftY, (int)width, (int)height, CORNER_SIZE);
+        RenderingHelper.renderDynamicTrim(sb, TEX_CORNER_TRIM, TEX_EDGE_TRIM, (int)bottomLeftX, (int)bottomLeftY, (int)width, (int)height, CORNER_SIZE, getTrimColor());
     }
 
     public void renderForeground(SpriteBatch sb, float bottomLeftX, float bottomLeftY, float width, float height) {
